xserver mesa git update 21 June 2013
xserver          commit 77e51d5bbb97eb5c9d9dbff9a7c44d7e53620e68
libxcb           commit 9ae84ad187e2ba440c40f44b8eb21c82c2fdbf12
libxcb/xcb-proto commit e5f7c750815cb5170db363a2e5b09639b7354733
xkeyboard-config commit 30d804538462213ed01e8efc0b44a8e5a0aff990
libX11           commit 9dfb0f3c0a761590bcdc1f3396b1e064da4e18e8
libXdmcp         commit 0b443c1b769b9c9a3b45b4252afe07e18b709ff4
libXext          commit d8366afbb0d2e4fbb1e419b1187f490522270bea
libfontenc       commit 3acba630d8b57084f7e92c15732408711ed5137a
libXinerama      commit 6e1d1dc328ba8162bba2f4694e7f3c706a1491ff
libXau           commit 899790011304c4029e15abf410e49ce7cec17e0a
xkbcomp          commit ed582f4fccd4e23abcfba8b3b03649fea6414f44
pixman           commit 279bdcda7ec3af8ac06312f4514b1b082a279544
xextproto        commit f27fcc99d1cf935cc289933326f7d3baacd5107a
randrproto       commit ca7cc541c2e43e6c784df19b4583ac35829d2f72
glproto          commit 8e3407e02980d088e20041e79bdcdd3737e7827e
mkfontscale      commit f731c5c36f28ddd0f25f474d2991c96f9a7a915c
xwininfo         commit ba0d1b0da21d2dbdd81098ed5778f3792b472e13
libXft           commit c5e760a239afc62a1c75e0509868e35957c8df52
libXmu           commit e46ecb4e02b7f919b11efa79448d4db71d1deb69
libxtrans        commit bd53f4c8543faf910a7a151241ee07661b4d57ad
fontconfig       commit cd9b1033a68816a7acfbba1718ba0aa5888f6ec7
mesa             commit bbd2d575e649c0929d125c7dd44723aeabcb2323

+++ xorg-server/hw/xwin/winmultiwindowicons.c
@@ -624,7 +624,7 @@ winDestroyIcon(HICON hIcon)
     /* Delete the icon if its not one of the application defaults or an override */
     if (hIcon &&
         hIcon != g_hIconX &&
-        hIcon != g_hSmallIconX && !winIconIsOverride((unsigned long) hIcon))
+        hIcon != g_hSmallIconX && !winIconIsOverride(hIcon))
         DestroyIcon(hIcon);
 }
 #endif
